Move node_library.standard to nodes.standard
And delete the (now empy) node_library module
